minor refactor (backend): resolve ImportError in evaluators and aggregation service
